+++ base/__init__.py
@@ -1,2 +1,2 @@
-VERSION = [1, 5, 2]
+VERSION = [1, 5, 3]
 __VERSION__ = '.'.join(map(str, VERSION))